requirements.txt
langchain
langchain-ollama
langchain-openai
langchain-community
langchain-chroma
python-dotenv
pydantic
pypdf
fastapi
uvicorn